@@ -27,7 +27,7 @@
     "@babel/preset-react": "^7.16.7",
     "@svgr/webpack": "^6.2.0",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^14.1.1",
     "babel-loader": "^8.2.3",
     "css-loader": "^6.5.1",
@@ -35,8 +35,8 @@
     "jest": "^28.0.0",
     "jest-environment-jsdom": "^28.0.0",
     "jsdom": "^19.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-transition-group": "^4.4.2",
     "sass": "^1.49.0",
     "sass-loader": "^12.4.0",
