@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import App from "router/App";
 import "./index.scss";
@@ -9,4 +10,6 @@ if (process.env.MODE !== "production") {
   reactAxe(React, ReactDOM, 1000);
 }
 
-ReactDOM.render(<App />, document.getElementById("app"));
+const container = document.getElementById("app");
+const root = createRoot(container);
+root.render(<App />);
